Fix labeling of magnitudes in table. Fix magnitude listed in header of list.
Magnitudes in tables on summary page are M 0-1, 1-2, etc not M>=1, M>=2,
etc. Use "M 0", "M 1", etc (dropping the "+").

Adding the mininum magnitude for historical earthquakes and aftershocks
means the earthquakes listed in the event list on the summary page may
be limited to larger magnitudes than the implemented default value
(based on mainshock magnitude). Added a fix to display the larger of the
minimum magnitude provided by the user and the default threshold for
listing so that that heading matches the actual data that appears.